refactor(profiles/api): change departments and categories profiles api
fix #912
