+++ src/api/bkuser_core/api/web/category/views.py
@@ -29,7 +29,6 @@ from .serializers import (
     CategoryMetaOutputSLZ,
     CategoryNamespaceSettingUpdateInputSLZ,
     CategoryProfileListInputSLZ,
-    CategoryProfileOutputSLZ,
     CategorySettingCreateInputSLZ,
     CategorySettingOutputSLZ,
     CategorySyncResponseOutputSLZ,
@@ -40,6 +39,7 @@ from .serializers import (
 from bkuser_core.api.web.department.serializers import DepartmentsWithChildrenAndAncestorsOutputSLZ
 from bkuser_core.api.web.export import ProfileExcelExporter
 from bkuser_core.api.web.field.serializers import FieldOutputSLZ
+from bkuser_core.api.web.serializers import ProfileDetailOutputSLZ
 from bkuser_core.api.web.utils import get_category, get_operator, list_setting_metas
 from bkuser_core.api.web.viewset import CustomPagination
 from bkuser_core.audit.constants import OperationType
@@ -557,7 +557,7 @@ class CategoryOperationSwitchOrderApi(generics.UpdateAPIView):
 class CategoryProfileListApi(generics.ListAPIView):
     permission_classes = [ViewCategoryPermission]
     pagination_class = CustomPagination
-    serializer_class = CategoryProfileOutputSLZ
+    serializer_class = ProfileDetailOutputSLZ
 
     def get_queryset(self):
         slz = CategoryProfileListInputSLZ(data=self.request.query_params)
@@ -573,6 +573,11 @@ class CategoryProfileListApi(generics.ListAPIView):
         if keyword:
             # NOTE: 这里相对原来的差异, 抹掉了 id__icontains 的搜索
             queryset = queryset.filter(Q(username__icontains=keyword) | Q(display_name__icontains=keyword))
+
+        # TODO: 支持 recursive=false 查询未关联部门的用户列表?
+
+        # TODO: 增加fields, 控制字段返回 => 搜索接口只需要几个字段
+        # id/username/display_name
 
         # do prefetch
         queryset = queryset.prefetch_related("departments", "leader")
