+++ src/api/bkuser_core/api/web/serializers.py
@@ -10,10 +10,16 @@ specific language governing permissions and limitations under the License.
 """
 import base64
 import datetime
-from typing import List
+from typing import TYPE_CHECKING, List
 
+from django.conf import settings
 from django.utils.encoding import force_bytes, force_str
 from rest_framework import fields, serializers
+
+from .utils import expand_extra_fields, get_extras_with_default_values
+
+if TYPE_CHECKING:
+    from bkuser_core.profiles.models import Profile
 
 # 公共的slz, 可能可以挪到别的地方, 后续再看
 
@@ -59,3 +65,69 @@ class Base64OrPlainField(serializers.CharField):
         if is_base64(data):
             return force_str(base64.b64decode(data))
         return super().to_internal_value(data)
+
+
+class ProfileDetailDepartmentSerializer(serializers.Serializer):
+    id = serializers.IntegerField(required=False)
+    name = serializers.CharField(required=False)
+    full_name = serializers.CharField(required=False)
+    category_id = serializers.IntegerField(required=False)
+
+
+class ProfileDetailLeaderSerializer(serializers.Serializer):
+    id = serializers.IntegerField(required=False, read_only=True)
+    username = serializers.CharField()
+    display_name = serializers.CharField(read_only=True)
+
+
+class ProfileDetailOutputSLZ(serializers.Serializer):
+    # Q: 不需要返回所有字段吧
+    # A: 需要, 目前前端的交互式, 列表点击直接拿数据渲染表单, 变更提交, 没有再次获取
+    id = serializers.CharField(required=False, help_text="用户ID")
+    username = serializers.CharField(required=False, help_text="用户名")
+    qq = serializers.CharField(required=False, help_text="QQ")
+    email = serializers.CharField(required=False, help_text="邮箱")
+    telephone = serializers.CharField(required=False, help_text="电话")
+    wx_userid = serializers.CharField(required=False, help_text="微信用户id")
+    domain = serializers.CharField(required=False, help_text="域")
+    display_name = serializers.CharField(required=False, help_text="中文名")
+    status = serializers.CharField(required=False, help_text="账户状态")
+    staff_status = serializers.CharField(required=False, help_text="在职状态")
+    position = serializers.CharField(required=False, help_text="职位")
+    enabled = serializers.BooleanField(required=False, help_text="是否启用", default=True)
+    # extras = serializers.JSONField(required=False, help_text="扩展字段")
+    password_valid_days = serializers.IntegerField(required=False, help_text="密码有效期")
+    account_expiration_date = serializers.CharField(required=False)
+    country_code = serializers.CharField(required=False, help_text="国家码")
+    iso_code = serializers.CharField(required=False, help_text="国家码")
+    time_zone = serializers.CharField(required=False, help_text="时区")
+
+    last_login_time = serializers.DateTimeField(required=False, help_text="最后登录时间")
+
+    create_time = serializers.DateTimeField(required=False, help_text="创建时间")
+    update_time = serializers.DateTimeField(required=False, help_text="更新时间")
+    departments = ProfileDetailDepartmentSerializer(many=True, required=False, help_text="部门列表")
+
+    # NOTE: 老的代码用的leader, 理论上应该换成leaders
+    # leaders = ProfileDetailLeaderSerializer(many=True, required=False, help_text="上级列表", source="leader")
+    leader = ProfileDetailLeaderSerializer(many=True, required=False, help_text="上级列表")
+
+    logo = serializers.SerializerMethodField(required=False)
+    extras = serializers.SerializerMethodField(required=False, read_only=True)
+
+    def get_extras(self, obj: "Profile") -> dict:
+        """尝试从 context 中获取默认字段值"""
+        return get_extras_with_default_values(obj.extras)
+
+    def get_logo(self, data):
+        logo = data.logo
+        if not logo:
+            # logo转成 "data:image/png;base64,xxxxx"
+            return settings.DEFAULT_LOGO_DATA
+
+        return logo
+
+    # NOTE: 部门下的用户列表需要把字段extras打平到profile的字段(用于页面展示+修改表单直接编辑/提交)
+    def to_representation(self, instance):
+        data = super().to_representation(instance)
+        return expand_extra_fields(data)
